Moving map test functions to the map class

+++ src/vector_map.h
@@ -23,7 +23,7 @@ template <typename K, typename V> struct vector_map {
   bool contains_key(K const &key) const;
 
 private:
-  int find(K const & key) const;
+  long long find(K const & key) const;
 
   struct Entry {
       K key;
@@ -67,7 +67,7 @@ size_t vector_map<K, V>::size() const {
 }
 
 template <typename K, typename V>
-int vector_map<K, V>::find(K const &key) const {
+long long vector_map<K, V>::find(K const &key) const {
   for (size_t i = 0; i < size(); ++i) {
     if (vector_[i].key == key) return i;
   }
@@ -105,7 +105,7 @@ bool vector_map<K, V>::add(K const &key, V value) {
 
 template <typename K, typename V>
 bool vector_map<K, V>::remove(K const &key) {
-  int index = find(key);
+  long long index = find(key);
   if (index == -1) return false;
   vector_.swap_remove(index);
   return true;
